Fixed VS2015 projects and cmake

+++ SRC/Common/ThreadWindows.cpp
@@ -58,5 +58,5 @@ void Thread::Suspend()
 
 void Thread::Sleep(size_t milliseconds)
 {
-	::Sleep(milliseconds);
+	::Sleep((DWORD)milliseconds);
 }
